src/cliente/requirements.txt
fastapi
uvicorn
jinja2
grpcio
grpcio-tools